IDVA6-466 Code refactored making file and function names follow agreed naming convention and to remove unnecessary imports
