+++ src/routers/controllers/companyAddedController.ts
@@ -1,0 +1,8 @@
+import { Request, Response, NextFunction } from "express";
+import { CompanyAddSuccessHandler } from "../handlers/yourCompanies/companyAddSuccessHandler";
+import * as constants from "../../constants";
+
+export const companyAddedControllerGet = async (req: Request, res: Response, next: NextFunction) => {
+    const viewData = await new CompanyAddSuccessHandler().execute(req, res);
+    res.render(constants.COMPANY_ADDED_SUCCESS_PAGE_TEMPLATE, viewData);
+};
