+++ src/routers/handlers/yourCompanies/addCompanyHandler.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { GenericHandler } from "../generic";
+import { GenericHandler } from "../genericHandler";
 import logger from "../../../lib/Logger";
 import {
     COMPANY_PROFILE,
@@ -11,7 +11,6 @@ import {
     ADD_COMPANY_LANG,
     COMPNANY_ASSOCIATED_WITH_USER,
     COMPANY_NUMBER,
-    ADD_COMPANY_URL,
     ENTER_A_COMPANY_NUMBER
 } from "../../../constants";
 import { CompanyProfile } from "@companieshouse/api-sdk-node/dist/services/company-profile/types";
