+++ test/src/routers/controllers/comfirmCompanyDetailsController.test.ts
@@ -1,14 +1,14 @@
-import mocks from "../../mocks/all.middleware.mock";
+import mocks from "../../../mocks/all.middleware.mock";
 import { Session } from "@companieshouse/node-session-handler";
 
-import app from "../../../src/app";
+import app from "../../../../src/app";
 import supertest from "supertest";
 import { NextFunction, Request, Response } from "express";
-import { validDisolvedCompanyProfile, validActiveCompanyProfile } from "../../mocks/companyProfile.mock";
+import { validDisolvedCompanyProfile, validActiveCompanyProfile } from "../../../mocks/companyProfile.mock";
 
 const router = supertest(app);
-const en = require("../../../src/locales/en/translation/confirm-company-details.json");
-const cy = require("../../../src/locales/cy/translation/confirm-company-details.json");
+const en = require("../../../../src/locales/en/translation/confirm-company-details.json");
+const cy = require("../../../../src/locales/cy/translation/confirm-company-details.json");
 const url = "/your-companies/confirm-company-details";
 
 const session: Session = new Session();
