+++ test/src/routers/controllers/comfirmationCompanyAddedController.test.ts
@@ -1,12 +1,12 @@
-import mocks from "../../mocks/all.middleware.mock";
-import app from "../../../src/app";
+import mocks from "../../../mocks/all.middleware.mock";
+import app from "../../../../src/app";
 import supertest from "supertest";
-import { getCompanyProfile } from "../../../src/services/companyProfileService";
-import { validActiveCompanyProfile } from "../../mocks/companyProfile.mock";
-jest.mock("../../../src/services/companyProfileService");
+import { getCompanyProfile } from "../../../../src/services/companyProfileService";
+import { validActiveCompanyProfile } from "../../../mocks/companyProfile.mock";
+jest.mock("../../../../src/services/companyProfileService");
 const router = supertest(app);
-const en = require("../../../src/locales/en/translation/confirmation-company-added.json");
-const cy = require("../../../src/locales/cy/translation/confirmation-company-added.json");
+const en = require("../../../../src/locales/en/translation/confirmation-company-added.json");
+const cy = require("../../../../src/locales/cy/translation/confirmation-company-added.json");
 const companyNumber = "12345678";
 const url = `/your-companies/company/${companyNumber}/confirmation-company-added`;
 const mockGetCompanyProfile = getCompanyProfile as jest.Mock;
